chore: 메서드 리네이밍, 위치 변경
