+++ src/main/java/com/keunsori/keunsoriserver/domain/reservation/controller/ReservationController.java
@@ -1,6 +1,7 @@
 package com.keunsori.keunsoriserver.domain.reservation.controller;
 
 import com.keunsori.keunsoriserver.domain.admin.reservation.dto.response.DailyAvailableResponse;
+import com.keunsori.keunsoriserver.domain.admin.reservation.service.AdminReservationService;
 import jakarta.validation.Valid;
 
 import org.springframework.http.ResponseEntity;
@@ -30,6 +31,7 @@ import lombok.RequiredArgsConstructor;
 public class ReservationController {
 
     private final ReservationService reservationService;
+    private final AdminReservationService adminReservationService;
 
     @GetMapping("/list")
     public ResponseEntity<List<ReservationResponse>> findAllReservations(@RequestParam("month") String month) {
@@ -66,7 +68,7 @@ public class ReservationController {
     // 예약 신청 페이지 가능한 날짜 반환
     @GetMapping
     public ResponseEntity<List<DailyAvailableResponse>> findMonthlySchedule(@RequestParam("month") String month){
-        List<DailyAvailableResponse> responses = reservationService.findDailyAvailableByMonth(month);
+        List<DailyAvailableResponse> responses = adminReservationService.findDailyAvailableByMonth(month);
         return ResponseEntity.ok().body(responses);
     }
 }
